Make sphinx_graphql.graphiql its own plugin
